modules: add module global download variable "MODULE_xxx_DOWNLOAD_SUCCESSIVE_INTERVAL"
It is unset by default. But can be assigned to a delay (in seconds).
A forced wait will be made between download of two links of a same hoster.

Some hosters can behave nasty (force user to wait, declare link as dead, or sometime worst) when downloading
successively a bunch of links.

This new global will be used soon in plowdown.