openload_co: add upload support
Support:
- upload for anonymous and free accounts
- --folder option - choosing and creating folders with subfolders

Media files are autoconverted and they are limited to max 10 GiB size,
normal files are limited to max 1 GiB size. Extensions for media files
were taken arbitrary. Tests were carried out with a 1.2GiB media test
file, so it should work reliably.

There is possibility for remote uploads, but maybe in the future it will
be implemented.

Implements and closes #74.