openload: fix login (upstream)
Previously I used a login and a password to login on service and grab an
API login and API key. Now we cannot simply login on our account,
because the send a mail with a login key, which is required for login.

Users now have to get an API login and API key from their accounts and
pass them as an authentication data.

It also implies that now we cannot create a folder, because API doesn't
have such function (previously I do it through WWW).