mediafire (download): finally fixes the captcha issue
On mediafire we have two authorization methods. One is solvemedia, and
second is a graphic version of recaptcha. In graphic version of recapcha
we for example have to choose all images with numbers. There isn't
a normal version of recaptcha on mediafire. It is always a graphic
version. When we use a recaptcha key to generate a normal version and
we send an answer to the server, it is always rejected by the server
and count as a wrong captcha.

So instead of resolving a graphic version of recaptcha, just force to
resolve solvemedia. Even if on a page there isn't solvemedia, the server
accepts this authorization method. It simply works.

Finally fixes the captcha issue #80.